Extract MCP server into packages/mcp, thin route wrapper in app
