+++ packages/app/package.json
@@ -37,6 +37,7 @@
     "@radix-ui/react-tooltip": "^1.2.8",
     "@semianalysisai/inferencex-constants": "workspace:*",
     "@semianalysisai/inferencex-db": "workspace:*",
+    "@semianalysisai/inferencex-mcp": "workspace:*",
     "@shikijs/rehype": "^4.0.2",
     "@tailwindcss/typography": "^0.5.19",
     "@tanstack/react-query": "^5.95.2",
