+++ packages/mcp/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@semianalysisai/inferencex-mcp",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./server": "./src/server.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@modelcontextprotocol/sdk": "^1.28.0",
+    "@semianalysisai/inferencex-constants": "workspace:*",
+    "@semianalysisai/inferencex-db": "workspace:*",
+    "zod": "^4.3.6"
+  },
+  "devDependencies": {
+    "@types/node": "^25.5.0",
+    "typescript": "^6.0.2"
+  }
+}
